move loadingBar from gallery to root app component
