+++ src/components/PhotoFrame.tsx
@@ -23,6 +23,7 @@ import PublicCollectionDownloadManager from 'services/publicCollectionDownloadMa
 import { PublicCollectionGalleryContext } from 'utils/publicCollectionGallery';
 import { useRouter } from 'next/router';
 import EmptyScreen from './EmptyScreen';
+import { AppContext } from 'pages/_app';
 
 const Container = styled.div`
     display: block;
@@ -86,6 +87,7 @@ const PhotoFrame = ({
     const [fetching, setFetching] = useState<{ [k: number]: boolean }>({});
     const startTime = Date.now();
     const galleryContext = useContext(GalleryContext);
+    const appContext = useContext(AppContext);
     const publicCollectionGalleryContext = useContext(
         PublicCollectionGalleryContext
     );
@@ -481,7 +483,7 @@ const PhotoFrame = ({
                     const mergedURL = galleryContext.files.get(item.id);
                     urls = mergedURL.split(',');
                 } else {
-                    galleryContext.startLoading();
+                    appContext.startLoading();
                     if (
                         publicCollectionGalleryContext.accessedThroughSharedURL
                     ) {
@@ -494,7 +496,7 @@ const PhotoFrame = ({
                     } else {
                         urls = await DownloadManager.getFile(item, true);
                     }
-                    galleryContext.finishLoading();
+                    appContext.finishLoading();
                     const mergedURL = urls.join(',');
                     galleryContext.files.set(item.id, mergedURL);
                 }
@@ -546,7 +548,9 @@ const PhotoFrame = ({
                                 filteredData={filteredData}
                                 activeCollection={activeCollection}
                                 showAppDownloadBanner={
-                                    files.length < 30 && !isInSearchMode
+                                    files.length < 30 &&
+                                    !isInSearchMode &&
+                                    !galleryContext.isDeduplicating
                                 }
                                 resetFetching={resetFetching}
                             />
