+++ src/pages/gallery/index.tsx
@@ -15,7 +15,6 @@ import {
     deleteFromTrash,
 } from 'services/fileService';
 import styled from 'styled-components';
-import LoadingBar from 'react-top-loading-bar';
 import {
     syncCollections,
     getCollectionsAndTheirLatestFile,
@@ -135,8 +134,7 @@ const defaultGalleryContext: GalleryContextType = {
     setActiveCollection: () => null,
     syncWithRemote: () => null,
     setDialogMessage: () => null,
-    startLoading: () => null,
-    finishLoading: () => null,
+
     setNotificationAttributes: () => null,
     setBlockingLoad: () => null,
     clubSameTimeFilesOnly: false,
@@ -194,10 +192,8 @@ export default function Gallery() {
         disabled: uploadInProgress,
     });
 
-    const loadingBar = useRef(null);
     const [isInSearchMode, setIsInSearchMode] = useState(false);
     const [searchStats, setSearchStats] = useState(null);
-    const isLoadingBarRunning = useRef(false);
     const syncInProgress = useRef(true);
     const resync = useRef(false);
     const [deleted, setDeleted] = useState<number[]>([]);
@@ -258,7 +254,7 @@ export default function Gallery() {
     useEffect(() => {
         const main = async () => {
             if (isDeduplicating) {
-                startLoading();
+                appContext.startLoading();
                 let duplicates = await getDuplicateFiles();
                 if (clubSameTimeFilesOnly) {
                     duplicates = clubDuplicatesByTime(duplicates);
@@ -295,7 +291,7 @@ export default function Gallery() {
                 }
                 setSelected(selectedFiles);
                 setActiveCollection(ALL_SECTION);
-                finishLoading();
+                appContext.finishLoading();
             } else {
                 setDuplicateFiles([]);
                 setFileSizeMap(new Map<number, number>());
@@ -363,7 +359,7 @@ export default function Gallery() {
             if (!(await isTokenValid())) {
                 throw new Error(ServerErrorCodes.SESSION_EXPIRED);
             }
-            !silent && startLoading();
+            !silent && appContext.startLoading();
             await billingService.syncSubscription();
             const collections = await syncCollections();
             setCollections(collections);
@@ -394,7 +390,7 @@ export default function Gallery() {
                     break;
             }
         } finally {
-            !silent && finishLoading();
+            !silent && appContext.finishLoading();
         }
         syncInProgress.current = false;
         if (resync.current) {
@@ -428,21 +424,12 @@ export default function Gallery() {
         setSelected({ count: 0, collectionID: 0 });
     };
 
-    const startLoading = () => {
-        !isLoadingBarRunning.current && loadingBar.current?.continuousStart();
-        isLoadingBarRunning.current = true;
-    };
-    const finishLoading = () => {
-        isLoadingBarRunning.current && loadingBar.current?.complete();
-        isLoadingBarRunning.current = false;
-    };
-
     if (!files) {
         return <div />;
     }
     const collectionOpsHelper =
         (ops: COLLECTION_OPS_TYPE) => async (collection: Collection) => {
-            startLoading();
+            appContext.startLoading();
             try {
                 await handleCollectionOps(
                     ops,
@@ -463,14 +450,14 @@ export default function Gallery() {
                 });
             } finally {
                 await syncWithRemote(false, true);
-                finishLoading();
+                appContext.finishLoading();
             }
         };
 
     const changeFilesVisibilityHelper = async (
         visibility: VISIBILITY_STATE
     ) => {
-        startLoading();
+        appContext.startLoading();
         try {
             const updatedFiles = await changeFilesVisibility(
                 files,
@@ -499,7 +486,7 @@ export default function Gallery() {
             });
         } finally {
             await syncWithRemote(false, true);
-            finishLoading();
+            appContext.finishLoading();
         }
     };
 
@@ -533,7 +520,7 @@ export default function Gallery() {
     };
 
     const deleteFileHelper = async (permanent?: boolean) => {
-        startLoading();
+        appContext.startLoading();
         try {
             const selectedFiles = getSelectedFiles(selected, files);
             if (permanent) {
@@ -564,7 +551,7 @@ export default function Gallery() {
             });
         } finally {
             await syncWithRemote(false, true);
-            finishLoading();
+            appContext.finishLoading();
         }
     };
 
@@ -594,7 +581,7 @@ export default function Gallery() {
             close: { text: constants.CANCEL },
         });
     const emptyTrashHelper = async () => {
-        startLoading();
+        appContext.startLoading();
         try {
             await emptyTrash();
             if (selected.collectionID === TRASH_SECTION) {
@@ -611,7 +598,7 @@ export default function Gallery() {
             });
         } finally {
             await syncWithRemote(false, true);
-            finishLoading();
+            appContext.finishLoading();
         }
     };
 
@@ -624,9 +611,9 @@ export default function Gallery() {
     const downloadHelper = async () => {
         const selectedFiles = getSelectedFiles(selected, files);
         clearSelection();
-        startLoading();
+        appContext.startLoading();
         await downloadFiles(selectedFiles);
-        finishLoading();
+        appContext.finishLoading();
     };
 
     return (
@@ -638,8 +625,7 @@ export default function Gallery() {
                 setActiveCollection,
                 syncWithRemote,
                 setDialogMessage,
-                startLoading,
-                finishLoading,
+
                 setNotificationAttributes,
                 setBlockingLoad,
                 clubSameTimeFilesOnly,
@@ -651,7 +637,6 @@ export default function Gallery() {
             <FullScreenDropZone
                 getRootProps={getRootProps}
                 getInputProps={getInputProps}>
-                <LoadingBar color="#51cd7c" ref={loadingBar} />
                 <AlertBanner bannerMessage={bannerMessage} />
                 <ToastNotification
                     attributes={notificationAttributes}
@@ -711,8 +696,6 @@ export default function Gallery() {
                             setCollectionNamerAttributes={
                                 setCollectionNamerAttributes
                             }
-                            startLoading={startLoading}
-                            finishLoading={finishLoading}
                             collectionFilesCount={collectionFilesCount}
                         />
                     </>
